refactor: extract abstract class and interface from AuthenticationInterceptor classes
