+++ src/main/java/nextstep/auth/authentication/AuthenticationConverter.java
@@ -1,0 +1,9 @@
+package nextstep.auth.authentication;
+
+import javax.servlet.http.HttpServletRequest;
+import java.io.IOException;
+
+public interface AuthenticationConverter {
+
+    AuthenticationToken convert(HttpServletRequest request) throws IOException;
+}
